fix fun

+++ app/src/main/java/com/jackpan/stockcomputer/viewmodel/MemberModel.kt
@@ -2,6 +2,7 @@ package com.jackpan.stockcomputer.viewmodel
 import android.app.Activity
 import android.app.Application
 import android.util.Log
+import android.widget.Toast
 import androidx.hilt.lifecycle.ViewModelInject
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.MutableLiveData
@@ -20,21 +21,27 @@ import com.google.firebase.ktx.Firebase
 import com.jackpan.stockcomputer.model.DeleteUserData
 import com.jackpan.stockcomputer.model.UserData
 import manager.TimeManager
+import manager.checkIsNull
 
-class MemberModel @ViewModelInject constructor(application: Application) : AndroidViewModel(application){
+class MemberModel @ViewModelInject constructor(application: Application, private val activity: Activity) : AndroidViewModel(application){
 
     val getSMSCode  = MutableLiveData<GetSMSCode>()
     val useData  = MutableLiveData<UserData>()
     var verificationIdString : String = ""
     val deleteUserData = MutableLiveData<DeleteUserData>()
+    lateinit var auth: FirebaseAuth
+    var optCode : String = ""
+    var phoneNumber : String = ""
+    var isCheck : Boolean = false
 
 
+    fun getPhoneSMScode(phoneNumber :String) = viewModelScope.launch {
 
-
-    fun getPhoneSMScode(auth:FirebaseAuth, phoneNumber :String, activity: Activity) = viewModelScope.launch {
+        auth = FirebaseAuth.getInstance()
+        auth.setLanguageCode("zh-TW")
 
         val options = PhoneAuthOptions.newBuilder(auth)
-            .setPhoneNumber(phoneNumber)       // Phone number to verify
+            .setPhoneNumber( phoneNumber.replaceFirst("0","+886"))       // Phone number to verify
             .setTimeout(60L, TimeUnit.SECONDS) // Timeout and unit
             .setActivity(activity)                 // Activity (for callback binding)
             .setCallbacks(object  :PhoneAuthProvider.OnVerificationStateChangedCallbacks(){
@@ -60,13 +67,13 @@ class MemberModel @ViewModelInject constructor(application: Application) : Andro
 
                 }
 
-            })          // OnVerificationStateChangedCallbacks
+            })
             .build()
         PhoneAuthProvider.verifyPhoneNumber(options)
     }
 
 
-    fun verifyPhoneNumberWithCode(auth:FirebaseAuth,activity: Activity, code: String){
+    fun verifyPhoneNumberWithCode(code: String){
         val credential = PhoneAuthProvider.getCredential(verificationIdString, code)
         auth.signInWithCredential(credential)
             .addOnCompleteListener(activity) { task ->
@@ -88,15 +95,33 @@ class MemberModel @ViewModelInject constructor(application: Application) : Andro
 
     }
 
-    fun  deleteUser(){
+    fun  deleteUser(isCheck : Boolean){
+        if (isCheck){
 
-        Firebase.auth.currentUser.delete().addOnCompleteListener {
-            if (it.isSuccessful){
-                deleteUserData.postValue(DeleteUserData(it.isSuccessful))
+            Firebase.auth.currentUser.delete().addOnCompleteListener {
+                if (it.isSuccessful){
+                    deleteUserData.postValue(DeleteUserData(it.isSuccessful))
+                }
+
+
             }
-
+        }else{
+            Toast.makeText(activity,"請先勾選同意刪除",Toast.LENGTH_SHORT).show()
 
         }
+
+    }
+    fun checkUserLogin(){
+
+        if (useData == null){
+            val user = Firebase.auth.currentUser
+         if (user!=null){
+            useData.postValue(UserData(user?.phoneNumber, TimeManager.getDate(user?.metadata?.creationTimestamp), TimeManager.getDate(user?.metadata?.lastSignInTimestamp), user?.uid))
+
+        }
+        }
+
+
     }
